+++ lib/common.dart
@@ -1,4 +1,5 @@
 library illuminate;
 
+export 'common/src/classes.dart' show Tuple, Coordinates;
 export 'common/src/helpers.dart' show tryCast, tryCastList, openExternally;
 export 'common/src/extensions.dart' show IterableExtension, StringExtension, ListExtension, ColorExtension;